refactor(NumericUpDown): split the control from the type it holds
Everything the control does apart from holding a number moves into
NumericUpDownBase: the template parts, the buttons and their repeat behaviour,
the keyboard and wheel handling, and the thirty odd properties that describe how
it looks rather than what it holds. NumericUpDown is what is left, and it is
still the double control it always was.

Nothing changes for anyone using it. The template still targets NumericUpDown,
every property sits where it did through inheritance, and the automation peer
takes the base now, so a control built on it reports its own name instead of
announcing that it is a NumericUpDown.
